Hide Help Updates
Moved feature to sauce. Fixed a bug where the user’s hide help setting
wasn’t saved between YNAB refreshes.
