--- source/common/res/features/hide-help/settings.json
@@ -1,14 +1,0 @@
-{
-         "name": "hideHelp",
-         "type": "checkbox",
-      "default": false,
-      "section": "general",
-        "title": "Hide Help (?) Button",
-  "description": "This feature hides the blue help (?) button in the bottom right corner of the screen. View the account-options popup (click your e-mail in the bottom left) to show or hide the help button.",
-      "actions": {
-                    "true": [
-                      "injectScript", "main.js",
-                      "injectCSS", "main.css"
-                    ]
-                 }
-}
